requirements.txt
pandas==2.0.3
pandera==0.17.0
minio==7.1.15
streamlit==1.24.1
protobuf>=3.20,<5
pyyaml==6.0.1
requests==2.31.0
pytest==7.4.4
mock==5.1.0
multimethod>=1.11.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
